MATH513 (MDSD5 2001A, MDSE2 2001A) {UOPMDISFEB2021}	R Documentation

Visualizing 2012 Presidential Dialogue

Description

Dataset of the three 2012 presidential debate between Obama and Romney, hosted by Schieffer, Crowley, and Lehrer with some questions from the crowd.

Format

Sample Data Sets as well as data frame prerequisite for package:
Data.frame with 2912 rows and 3 columns:
Person: speaker name
Dialogue: dialogue from the speaker
Turn: Integer field for dialogue sequence

Note

Instructions on package installation available on https://github.com/zeonlee/math513
Package files avalable on https://github.com/zeonlee/math513files

Author(s)

Cassandria Goh (10715753) Lim Jun Yong (10715755) Muhammad Firdaus (10715756) Lee Wei Jian (10715754)

Maintainer: The package maintainer <zeon.sg@gmail.com>

References

"10 Things You Didn't Know About Steven Chu." U.S. News & World Report, U.S. News & World Report, www.usnews.com/news/obama/articles/2008/12/30/10-things-you-didnt-know-about-steven-chu.

"October 16, 2012 Debate Transcript." CPD: October 16, 2012 Debate Transcript, www.debates.org/voter-education/debate-transcripts/october-16-2012-the-second-obama-romney-presidential-debate/.

"October 22, 2012 Debate Transcript." CPD: October 22, 2012 Debate Transcript, www.debates.org/voter-education/debate-transcripts/october-22-2012-the-third-obama-romney-presidential-debate/.

"October 3, 2012 Debate Transcript." CPD: October 3, 2012 Debate Transcript, www.debates.org/voter-education/debate-transcripts/october-3-2012-debate-transcript/

[Package UOPMDISFEB2021 version 0.1.0 Index]