filterwordsdf {UOPMDISFEB2021}	R Documentation

Filtering dataframe to display words that are entered in a vector form

Description

This function will filter the data frame's words according to those that are in vec

Arguments

data
data frame

vec
vector of strings

Value

data frame

Examples


 ## Not run: 
 filterwordsdf(dataframe, vector)

## End(Not run)

[Package UOPMDISFEB2021 version 0.1.0 Index]