mostwords {UOPMDISFEB2021}	R Documentation

Finding the most words in data

Description

Find top m words and Sort these words according to number of times appeared and only return top m

Arguments

data
mostwords(data, n)

Value

data frame

Examples


## Not run: 
 mostwords(data, n)

## End(Not run)

[Package UOPMDISFEB2021 version 0.1.0 Index]