persontop {UOPMDISFEB2021}	R Documentation

Group the result by person and find their top n most used words

Description

Group the result by person and find their top n most used words

Arguments

data
(n)

Value

persontop(n = n)

Examples


## Not run: 
 persontop(n)

## End(Not run)

[Package UOPMDISFEB2021 version 0.1.0 Index]